Fixed sizing issues

+++ HyperscanningNetworkLogger.h
@@ -43,7 +43,7 @@ class HyperscanningNetworkLogger : public EnvironmentExtension, public Thread {
 		bool mLogNetwork;
 
 		std::vector<std::string> mSharedStates;
-		std::vector<uint32_t> mStateValues;
+		std::vector<uint64_t> mStateValues;
 		std::vector<bool> mHasUpdated;
 		std::mutex mStateValuesMutex;
 		std::string mMessage;
